Refactoring tests Refactored out the logic that computes whether a hole array is resolved.
